feat: change ui component to tdesign; remove the concept of current cluster

+++ pkg/web/src/components/overview/OverviewTablePanel.tsx
@@ -1,69 +1,77 @@
-import { Alert, Button, Card, Modal, Table, TableColumn, Text } from 'tea-component';
 import { v4 } from 'uuid';
 
-import React, { memo, useEffect, useMemo, useState } from 'react';
+import React, { memo, useMemo } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
+import { Alert, Button, Dialog, Divider, PrimaryTableCol, Table } from 'tdesign-react';
 
 import { clusterApi } from '../../apis/clusterApi';
 import { useSelector } from '../../hooks';
 import { ClusterSimpleInfo } from '../../models';
 import { editClusterActions } from '../../store/editClusterSlice';
+import { insightAction } from '../../store/insightSlice';
 import { getErrorMsg } from '../../utils/getErrorMsg';
+import { Card } from '../common/Card';
 
 export const OverviewTablePanel = memo(() => {
   const { t } = useTranslation();
   const dispatch = useDispatch();
+  const [pagination, setPagination] = React.useState<{ current: number; pageSize: number }>({
+    current: 1,
+    pageSize: 10
+  });
   const [deleteDialog, setDeleteDialog] = React.useState<{
     visible: boolean;
     clusterName?: string;
     clusterId?: string;
   }>({ visible: false });
 
-  const searchFilter = useSelector(state => state.overview.searchFilter);
+  const selectedClusterId = useSelector(state => state.insight.selectedClusterId);
+  const searchText = useSelector(state => state.overview.searchText);
 
   const clusterList = clusterApi.useFetchClusterListQuery();
   const [deleteCluster, deleteClusterOptions] = clusterApi.useDeleteClusterMutation();
 
   const columns = useMemo(() => {
-    const columns: TableColumn<ClusterSimpleInfo>[] = [
+    const columns: PrimaryTableCol<ClusterSimpleInfo>[] = [
       {
-        key: 'id',
-        header: t('ID/名称'),
-        render: cluster => {
+        colKey: 'id',
+        title: t('ID/名称'),
+        width: '15%',
+        render: ({ row: cluster }) => {
           return (
             <div>
-              <Text parent="div">{cluster.id}</Text>
-              <Text parent="div" theme="label">
-                {cluster.name}
-              </Text>
+              <div>{cluster.id}</div>
+              <div>{cluster.name}</div>
             </div>
           );
         }
       },
       {
-        key: 'craneUrl',
-        header: t('Crane URL'),
-        render: cluster => {
+        colKey: 'craneUrl',
+        title: t('Crane URL'),
+        render: ({ row: cluster }) => {
           return cluster.craneUrl;
         }
       },
       {
-        key: 'operation',
-        header: t('操作'),
-        render: cluster => {
+        colKey: 'operation',
+        title: t('操作'),
+        render: ({ row: cluster }) => {
           return (
             <>
               <Button
-                type="link"
+                style={{ padding: 0 }}
+                theme="primary"
+                variant="text"
                 onClick={() => {
                   dispatch(editClusterActions.modalVisible(true));
                   dispatch(editClusterActions.mode('update'));
+                  dispatch(editClusterActions.editingClusterId(cluster.id));
                   dispatch(
                     editClusterActions.setClusters([
                       {
-                        id: v4(),
-                        clusterId: cluster.id,
+                        id: cluster.id,
                         clusterName: cluster.name,
                         craneUrl: cluster.craneUrl
                       }
@@ -73,8 +81,11 @@ export const OverviewTablePanel = memo(() => {
               >
                 {t('更新')}
               </Button>
+              <Divider layout="vertical" />
               <Button
-                type="link"
+                style={{ paddingLeft: '0.5rem', paddingRight: 0 }}
+                theme="primary"
+                variant="text"
                 onClick={() => {
                   setDeleteDialog({ visible: true, clusterId: cluster.id, clusterName: cluster.name });
                 }}
@@ -98,33 +109,35 @@ export const OverviewTablePanel = memo(() => {
 
   const filteredRecords = React.useMemo(() => {
     const list = clusterList.data?.data?.items ?? [];
-    return searchFilter && (searchFilter?.clusterIds?.length > 0 || searchFilter?.clusterNames?.length) > 0
+    return searchText
       ? list.filter(item => {
-          return (
-            (searchFilter?.clusterIds ?? []).includes(item.id) || (searchFilter?.clusterNames ?? []).includes(item.name)
-          );
+          return item.id === searchText || item.name === searchText;
         })
       : list;
-  }, [clusterList.data?.data?.items, searchFilter]);
+  }, [clusterList.data?.data?.items, searchText]);
 
   return (
     <>
-      <Modal caption={t('删除集群')} visible={deleteDialog.visible} onClose={() => setDeleteDialog({ visible: false })}>
-        <Modal.Body>
-          <Trans>
-            确认删除集群: {{ clusterName: deleteDialog.clusterName ?? '' }}
-            {deleteClusterOptions.isError && (
-              <Alert className="tea-mt-3n tea-mb-0" type="error">
-                {getErrorMsg(deleteClusterOptions.error)}
-              </Alert>
-            )}
-          </Trans>
-        </Modal.Body>
-        <Modal.Footer>
+      <Dialog
+        cancelBtn={
+          <Button
+            theme="default"
+            variant="base"
+            onClick={() => {
+              setDeleteDialog({ visible: false });
+            }}
+          >
+            {t('取消')}
+          </Button>
+        }
+        confirmBtn={
           <Button
             loading={deleteClusterOptions.isLoading}
-            type="primary"
             onClick={() => {
+              // if it is deleting the selected cluster, deselect it
+              if (deleteDialog.clusterId === selectedClusterId) {
+                dispatch(insightAction.selectedClusterId(null));
+              }
               deleteCluster({
                 clusterId: deleteDialog.clusterId
               });
@@ -132,34 +145,41 @@ export const OverviewTablePanel = memo(() => {
           >
             {t('确定')}
           </Button>
-          <Button
-            type="weak"
-            onClick={() => {
-              setDeleteDialog({ visible: false });
-            }}
-          >
-            取消
-          </Button>
-        </Modal.Footer>
-      </Modal>
-      <Card>
-        <Card.Body>
-          <Table
-            addons={[
-              Table.addons.pageable(),
-              Table.addons.autotip({
-                isError: clusterList.isError,
-                onRetry: () => {
-                  clusterList.refetch();
-                },
-                isLoading: clusterList.isLoading
-              })
-            ]}
-            columns={columns}
-            recordKey={record => record.id}
-            records={filteredRecords}
-          />
-        </Card.Body>
+        }
+        header={t('删除集群')}
+        visible={deleteDialog.visible}
+        onClose={() => setDeleteDialog({ visible: false })}
+      >
+        <div>
+          {t('确认删除集群')}: {deleteDialog.clusterName ?? ''}
+          {deleteClusterOptions.isError && (
+            <Alert
+              message={getErrorMsg(deleteClusterOptions.error)}
+              style={{ marginBottom: 0, marginTop: '1rem' }}
+              theme="error"
+            />
+          )}
+        </div>
+      </Dialog>
+      <Card style={{ marginTop: '1rem' }}>
+        <Table
+          columns={columns}
+          data={filteredRecords}
+          empty={t('暂无数据')}
+          pagination={{
+            pageSizeOptions: [],
+            totalContent: t('共 {total} 项数据', { total: filteredRecords.length }),
+            pageSize: pagination.pageSize,
+            current: pagination.current,
+            onCurrentChange: current => {
+              setPagination(pagination => ({ ...pagination, current }));
+            },
+            onPageSizeChange: pageSize => {
+              setPagination(pagination => ({ ...pagination, pageSize }));
+            }
+          }}
+          rowKey={'id'}
+        />
       </Card>
     </>
   );
